Language.C.Analysis.DeclAnalysis

Description

This module performs the analysis of declarations and the translation of type specifications in the AST.

Translating types

analyseTypeDecl :: MonadTrav m => CDecl -> m Type

get the type of a type declaration

A type declaration T may appear in thre forms:

typeof(T)

as abstract declarator in a function prototype, as in f(int)

in a declaration without declarators, as in struct x { int a } ;

Currently, analyseTypeDecl is exlusively used for analysing types for GNU's typeof(T).

We move attributes to the type, as they have no meaning for the abstract declarator

tType :: MonadTrav m => Bool -> NodeInfo -> [CTypeQual] -> TypeSpecAnalysis -> [CDerivedDeclr] -> [CDecl] -> m Type

translate a type

tDirectType :: MonadTrav m => Bool -> NodeInfo -> [CTypeQual] -> TypeSpecAnalysis -> m Type

translate a type without (syntactic) indirections Due to the GNU typeof extension and typeDefs, this can be an arbitrary type

tNumType :: MonadCError m => NumTypeSpec -> m (Either (FloatType, Bool) IntType)

Mapping from num type specs to C types (C99 6.7.2-2), ignoring the complex qualifier.

tArraySize :: MonadTrav m => CArrSize -> m ArraySize

tTypeQuals :: MonadTrav m => [CTypeQual] -> m (TypeQuals, Attributes)

mergeOldStyle :: MonadCError m => NodeInfo -> [CDecl] -> [CDerivedDeclr] -> m [CDerivedDeclr]

convert old style parameters

This requires matching parameter names and declarations, as in the following example:

int f(d,c,a,b)
char a,*b;
int c;
{ }

is converted to

int f(int d, int c, char a, char* b)

TODO: This could be moved to syntax, as it operates on the AST only

Dissecting type specs

canonicalTypeSpec :: MonadTrav m => [CTypeSpec] -> m TypeSpecAnalysis

data NumBaseType

Constructors

NoBaseType

BaseChar

BaseInt

BaseInt128

BaseFloat

BaseFloat128

BaseDouble

data SignSpec

Constructors

NoSignSpec

Signed

Unsigned

data SizeMod

Constructors

NoSizeMod

ShortMod

LongMod

LongLongMod

data NumTypeSpec

Constructors

NumTypeSpec

Fields

base :: NumBaseType

signSpec :: SignSpec

sizeMod :: SizeMod

isComplex :: Bool

data TypeSpecAnalysis

Constructors

TSNone

TSVoid

TSBool

TSNum NumTypeSpec

TSTypeDef TypeDefRef

TSType Type

TSNonBasic CTypeSpec

canonicalStorageSpec :: MonadCError m => [CStorageSpec] -> m StorageSpec

data StorageSpec

Constructors

NoStorageSpec

AutoSpec

RegSpec

ThreadSpec

StaticSpec Bool

ExternSpec Bool

hasThreadLocalSpec :: StorageSpec -> Bool

isTypeDef :: [CDeclSpec] -> Bool

Helpers

data VarDeclInfo

Constructors

VarDeclInfo VarName FunctionAttrs StorageSpec Attributes Type NodeInfo

tAttr :: (MonadCError m, MonadSymtab m) => CAttr -> m Attr

translate attribute annotations TODO: This is a unwrap and wrap stub

mkVarName :: (MonadCError m, MonadSymtab m) => NodeInfo -> Maybe Ident -> Maybe AsmName -> m VarName

construct a name for a variable TODO: more or less bogus

getOnlyDeclr :: MonadCError m => CDecl -> m CDeclr

nameOfDecl :: MonadCError m => CDecl -> m Ident

analyseVarDecl :: MonadTrav m => Bool -> [CStorageSpec] -> [CAttr] -> [CTypeQual] -> TypeSpecAnalysis -> [CFunSpec] -> CDeclr -> [CDecl] -> Maybe CInit -> m VarDeclInfo

analyse declarators

analyseVarDecl' :: MonadTrav m => Bool -> [CDeclSpec] -> CDeclr -> [CDecl] -> Maybe CInit -> m VarDeclInfo