Language.C.Analysis.DefTable

Description

This module manages symbols in local and global scopes.

There are four different kind of identifiers: ordinary identifiers (henceforth simply called identifier), tag names (names of struct/union/enum types), labels and structure members.

Documentation

type IdentEntry = Either TypeDef IdentDecl

All ordinary identifiers map to IdenTyDecl: either a typedef or a object/function/enumerator

identOfTyDecl :: IdentEntry -> Ident

type TagEntry = Either TagFwdDecl TagDef

Tag names map to forward declarations or definitions of struct/union/enum types

data TagFwdDecl

Constructors

CompDecl CompTypeRef

EnumDecl EnumTypeRef

data DefTable

Table holding current definitions

Constructors

DefTable

Fields

identDecls :: NameSpaceMap Ident IdentEntry

declared `ordinary identifiers'

tagDecls :: NameSpaceMap SUERef TagEntry

declared structunionenum tags

labelDefs :: NameSpaceMap Ident Ident

defined labels

memberDecls :: NameSpaceMap Ident MemberDecl

member declarations (only local)

refTable :: IntMap Name

link names with definitions

typeTable :: IntMap Type

emptyDefTable :: DefTable

empty definition table, with all name space maps in global scope

globalDefs :: DefTable -> GlobalDecls

get the globally defined entries of a definition table

inFileScope :: DefTable -> Bool

enterFunctionScope :: DefTable -> DefTable

Enter function scope (AND the corresponding block scope)

leaveFunctionScope :: DefTable -> DefTable

Leave function scope, and return the associated DefTable. Error if not in function scope.

enterBlockScope :: DefTable -> DefTable

Enter new block scope

leaveBlockScope :: DefTable -> DefTable

Leave innermost block scope

enterMemberDecl :: DefTable -> DefTable

Enter new member declaration scope

leaveMemberDecl :: DefTable -> ([MemberDecl], DefTable)

Leave innermost member declaration scope

data DeclarationStatus t

Status of a declaration

Constructors

NewDecl
new entry

Redeclared t
old def was overwritten

KeepDef t
new def was discarded

Shadowed t
new def shadows one in outer scope

KindMismatch t
kind mismatch

declStatusDescr :: DeclarationStatus t -> String

defineTypeDef :: Ident -> TypeDef -> DefTable -> (DeclarationStatus IdentEntry, DefTable)

defineGlobalIdent :: Ident -> IdentDecl -> DefTable -> (DeclarationStatus IdentEntry, DefTable)

declare/define a global object/function/typeDef

returns Redeclared def if there is already an object/function/typeDef in global scope, or DifferentKindRedec def if the old declaration is of a different kind.

defineScopedIdent :: Ident -> IdentDecl -> DefTable -> (DeclarationStatus IdentEntry, DefTable)

declare/define a object/function/typeDef with lexical scope

returns Redeclared def or DifferentKindRedec def if there is already an object/function/typeDef in the same scope.

defineScopedIdentWhen :: (IdentDecl -> Bool) -> Ident -> IdentDecl -> DefTable -> (DeclarationStatus IdentEntry, DefTable)

declare/define a object/function/typeDef with lexical scope, if the given predicate holds on the old entry.

returns Keep old_def if the old definition shouldn't be overwritten, and otherwise Redeclared def or DifferentKindRedecl def if there is already an object/function/typeDef in the same scope.

declareTag :: SUERef -> TagFwdDecl -> DefTable -> (DeclarationStatus TagEntry, DefTable)

declare a tag (fwd decl in case the struct name isn't defined yet)

defineTag :: SUERef -> TagDef -> DefTable -> (DeclarationStatus TagEntry, DefTable)

define a tag

defineLabel :: Ident -> DefTable -> (DeclarationStatus Ident, DefTable)

define a label Return the old label if it is already defined in this function's scope

lookupIdent :: Ident -> DefTable -> Maybe IdentEntry

lookup identifier (object, function, typeDef, enumerator)

lookupTag :: SUERef -> DefTable -> Maybe TagEntry

lookup tag

lookupLabel :: Ident -> DefTable -> Maybe Ident

lookup label

lookupIdentInner :: Ident -> DefTable -> Maybe IdentEntry

lookup an object in the innermost scope

lookupTagInner :: SUERef -> DefTable -> Maybe TagEntry

lookup an identifier in the innermost scope

insertType :: DefTable -> Name -> Type -> DefTable

Record the type of a node.

lookupType :: DefTable -> Name -> Maybe Type

Lookup the type of a node.

mergeDefTable :: DefTable -> DefTable -> DefTable

Merge two DefTables. If both tables contain an entry for a given key, they must agree on its value.