Language.C.Analysis.SemError

Description

Errors in the semantic analysis

Documentation

newtype InvalidASTError

InvalidASTError is caused by the violation of an invariant in the AST

Constructors

InvalidAST ErrorInfo

invalidAST :: NodeInfo -> String -> InvalidASTError

newtype BadSpecifierError

BadSpecifierError is caused by an invalid combination of specifiers

Constructors

BadSpecifierError ErrorInfo

badSpecifierError :: NodeInfo -> String -> BadSpecifierError

data TypeMismatch

Constructors

TypeMismatch String (NodeInfo, Type) (NodeInfo, Type)

typeMismatch :: String -> (NodeInfo, Type) -> (NodeInfo, Type) -> TypeMismatch

data RedefError

RedefError is caused by an invalid redefinition of the same identifier or type

Constructors

RedefError ErrorLevel RedefInfo

data RedefInfo

Constructors

RedefInfo String RedefKind NodeInfo NodeInfo

data RedefKind

Constructors

DuplicateDef

DiffKindRedecl

ShadowedDef

DisagreeLinkage

NoLinkageOld

redefinition :: ErrorLevel -> String -> RedefKind -> NodeInfo -> NodeInfo -> RedefError