Language.C.Analysis.TypeConversions

Documentation

arithmeticConversion :: TypeName -> TypeName -> Maybe TypeName

For an arithmetic operator, if the arguments are of the given types, return the type of the full expression.

floatConversion :: FloatType -> FloatType -> FloatType

intConversion :: IntType -> IntType -> IntType