Language.C.Data.Ident

Description

This module provides the notion of identifiers in C, speed up using hashing. Identifiers are associated optionally associated with a NodeInfo, i.e. with a unique Name and a source location (Position). The ordering relation on identifiers is based on the hash and does not follow the lexical order.

Documentation

data Ident

C identifiers

Constructors

Ident String !Int NodeInfo

data SUERef

References uniquely determining a struct, union or enum type. Those are either identified by an string identifier, or by a unique name (anonymous types).

Constructors

AnonymousRef Name

NamedRef Ident

isAnonymousRef :: SUERef -> Bool

Return true if the struct/union/enum reference is anonymous.

mkIdent :: Position -> String -> Name -> Ident

build an identifier from a string.

only minimal error checking, e.g., the characters of the identifier are not checked for being alphanumerical only; the correct lexis of the identifier should be ensured by the caller, e.g., the scanner.

for reasons of simplicity the complete lexeme is hashed.

builtinIdent :: String -> Ident

returns a builtin identifier (has builtin position and no unique name)

internalIdent :: String -> Ident

returns an internal identifier (has internal position and no unique name)

internalIdentAt :: Position -> String -> Ident

return an internal identifier with position info

isInternalIdent :: Ident -> Bool

return True if the given identifier is internal

identToString :: Ident -> String

string of an identifier

sueRefToString :: SUERef -> String

string of a SUE ref (empty if anonymous)

dumpIdent :: Ident -> String

dump the identifier string and its positions for debugging purposes