Language.C.Data.Node

Description

source position and unqiue name

Documentation

data NodeInfo

Parsed entity attribute

Constructors

OnlyPos Position !PosLength

NodeInfo Position !PosLength !Name

undefNode :: NodeInfo

create a node with neither name nor positional information

isUndefNode :: NodeInfo -> Bool

return True if the node carries neither name nor positional information

mkNodeInfoOnlyPos :: Position -> NodeInfo

| Given only a source position, create a new node attribute

mkNodeInfoPosLen :: Position -> PosLength -> NodeInfo

Given a source position and the position and length of the last token, create a new node attribute

mkNodeInfo :: Position -> Name -> NodeInfo

Given a source position and a unique name, create a new attribute identifier

mkNodeInfo' :: Position -> PosLength -> Name -> NodeInfo

Given a source position, the position and length of the last token and a unique name, create a new attribute identifier. Strict in

internalNode :: NodeInfo

Deprecated: use undefNode instead

class CNode a where

a class for convenient access to the attributes of an attributed object

Minimal complete definition

nodeInfo

Methods

nodeInfo :: a -> NodeInfo

fileOfNode :: CNode a => a -> Maybe FilePath

posOfNode :: NodeInfo -> Position

nameOfNode :: NodeInfo -> Maybe Name

getLastTokenPos :: NodeInfo -> PosLength

get the position and length of the last token

lengthOfNode :: NodeInfo -> Maybe Int

get the number of characters an AST node spans

eqByName :: CNode a => a -> a -> Bool

equality by name