Language.C.Analysis.SemRep

Description

This module contains definitions for representing C translation units. In contrast to AST, the representation tries to express the semantics of of a translation unit.

Sums of tags and identifiers

data TagDef

Composite type definitions (tags)

Constructors

CompDef CompType

EnumDef EnumType

typeOfTagDef :: TagDef -> TypeName

return the type corresponding to a tag definition

class Declaration n where

All datatypes aggregating a declaration are instances of Declaration

Methods

getVarDecl :: n -> VarDecl

get the name, type and declaration attributes of a declaration or definition

declIdent :: Declaration n => n -> Ident

get the variable identifier of a declaration (only safe if the the declaration is known to have a name)

declName :: Declaration n => n -> VarName

get the variable name of a Declaration

declType :: Declaration n => n -> Type

get the type of a Declaration

declAttrs :: Declaration n => n -> DeclAttrs

get the declaration attributes of a Declaration

data IdentDecl

identifiers, typedefs and enumeration constants (namespace sum)

Constructors

Declaration Decl
object or function declaration

ObjectDef ObjDef
object definition

FunctionDef FunDef
function definition

EnumeratorDef Enumerator
definition of an enumerator

objKindDescr :: IdentDecl -> String

textual description of the kind of an object

splitIdentDecls :: Bool -> Map Ident IdentDecl -> (Map Ident Decl, (Map Ident Enumerator, Map Ident ObjDef, Map Ident FunDef))

splitIdentDecls includeAllDecls splits a map of object, function and enumerator declarations and definitions into one map holding declarations, and three maps for object definitions, enumerator definitions and function definitions. If includeAllDecls is True all declarations are present in the first map, otherwise only those where no corresponding definition is available.

Global definitions

data GlobalDecls

global declaration/definition table returned by the analysis

Constructors

GlobalDecls

Fields

gObjs :: Map Ident IdentDecl

gTags :: Map SUERef TagDef

gTypeDefs :: Map Ident TypeDef

emptyGlobalDecls :: GlobalDecls

empty global declaration table

filterGlobalDecls :: (DeclEvent -> Bool) -> GlobalDecls -> GlobalDecls

filter global declarations

mergeGlobalDecls :: GlobalDecls -> GlobalDecls -> GlobalDecls

merge global declarations

Events for visitors

data DeclEvent

Declaration events

Those events are reported to callbacks, which are executed during the traversal.

Constructors

TagEvent TagDef
file-scope struct/union/enum event

DeclEvent IdentDecl
file-scope declaration or definition

ParamEvent ParamDecl
parameter declaration

LocalEvent IdentDecl
local variable declaration or definition

TypeDefEvent TypeDef
a type definition

AsmEvent AsmBlock
assembler block

Declarations and definitions

data Decl

Declarations, which aren't definitions

Constructors

Decl VarDecl NodeInfo

data ObjDef

Object Definitions

An object definition is a declaration together with an initializer.

If the initializer is missing, it is a tentative definition, i.e. a definition which might be overriden later on.

Constructors

ObjDef VarDecl (Maybe Initializer) NodeInfo

isTentative :: ObjDef -> Bool

Returns True if the given object definition is tentative.

data FunDef

Function definitions

A function definition is a declaration together with a statement (the function body).

Constructors

FunDef VarDecl Stmt NodeInfo

data ParamDecl

Parameter declaration

Constructors

ParamDecl VarDecl NodeInfo

AbstractParamDecl VarDecl NodeInfo

data MemberDecl

Struct/Union member declaration

Constructors

MemberDecl VarDecl (Maybe Expr) NodeInfo
MemberDecl vardecl bitfieldsize node

AnonBitField Type Expr NodeInfo
AnonBitField typ size

data TypeDef

typedef definitions.

The identifier is a new name for the given type.

Constructors

TypeDef Ident Type Attributes NodeInfo

identOfTypeDef :: TypeDef -> Ident

return the idenitifier of a typedef

data VarDecl

Generic variable declarations

Constructors

VarDecl VarName DeclAttrs Type

Declaration attributes

data DeclAttrs

Declaration attributes of the form DeclAttrs isInlineFunction storage linkage attrs

They specify the storage and linkage of a declared object.

Constructors

DeclAttrs Bool Storage Attributes
DeclAttrs inline storage attrs

isExtDecl :: Declaration n => n -> Bool

data Storage

Storage duration and linkage of a variable

Constructors

NoStorage
no storage

Auto Register
automatic storage (optional: register)

Static Linkage ThreadLocal
static storage, linkage spec and thread local specifier (gnu c)

FunLinkage Linkage
function, either internal or external linkage

declStorage :: Declaration d => d -> Storage

get the Storage of a declaration

type ThreadLocal = Bool

type Register = Bool

data Linkage

Linkage: Either no linkage, internal to the translation unit or external

Constructors

NoLinkage

InternalLinkage

ExternalLinkage

hasLinkage :: Storage -> Bool

return True if the object has linkage

declLinkage :: Declaration d => d -> Linkage

Get the linkage of a definition

Types

data Type

types of C objects

Constructors

DirectType TypeName TypeQuals Attributes
a non-derived type

PtrType Type TypeQuals Attributes
pointer type

ArrayType Type ArraySize TypeQuals Attributes
array type

FunctionType FunType Attributes
function type

TypeDefType TypeDefRef TypeQuals Attributes
a defined type

data FunType

Function types are of the form FunType return-type params isVariadic.

If the parameter types aren't yet known, the function has type FunTypeIncomplete type attrs.

Constructors

FunType Type [ParamDecl] Bool

FunTypeIncomplete Type

data ArraySize

An array type may either have unknown size or a specified array size, the latter either variable or constant. Furthermore, when used as a function parameters, the size may be qualified as static. In a function prototype, the size may be `Unspecified variable size' ([*]).

Constructors

UnknownArraySize Bool
UnknownArraySize is-starred

ArraySize Bool Expr
FixedSizeArray is-static size-expr

data TypeDefRef

typdef references If the actual type is known, it is attached for convenience

Constructors

TypeDefRef Ident (Maybe Type) NodeInfo

data TypeName

normalized type representation

Constructors

TyVoid

TyIntegral IntType

TyFloating FloatType

TyComplex FloatType

TyComp CompTypeRef

TyEnum EnumTypeRef

TyBuiltin BuiltinType

data BuiltinType

Builtin type (va_list, anything)

Constructors

TyVaList

TyAny

data IntType

integral types (C99 6.7.2.2)

Constructors

TyBool

TyChar

TySChar

TyUChar

TyShort

TyUShort

TyInt

TyUInt

TyLong

TyULong

TyLLong

TyULLong

data FloatType

floating point type (C99 6.7.2.2)

Constructors

TyFloat

TyDouble

TyLDouble

class HasSUERef a where

accessor class : struct/union/enum names

Methods

sueRef :: a -> SUERef

class HasCompTyKind a where

accessor class : composite type tags (struct or union)

Methods

compTag :: a -> CompTyKind

data CompTypeRef

composite type declarations

Constructors

CompTypeRef SUERef CompTyKind NodeInfo

data CompType

Composite type (struct or union).

Constructors

CompType SUERef CompTyKind [MemberDecl] Attributes NodeInfo

typeOfCompDef :: CompType -> TypeName

return the type of a composite type definition

data CompTyKind

a tag to determine wheter we refer to a struct or union, see CompType.

Constructors

StructTag

UnionTag

data EnumTypeRef

Constructors

EnumTypeRef SUERef NodeInfo

data EnumType

Representation of C enumeration types

Constructors

EnumType SUERef [Enumerator] Attributes NodeInfo
EnumType name enumeration-constants attrs node

typeOfEnumDef :: EnumType -> TypeName

return the type of an enum definition

data Enumerator

An Enumerator consists of an identifier, a constant expressions and the link to its type

Constructors

Enumerator Ident Expr EnumType NodeInfo

data TypeQuals

Type qualifiers: constant, volatile and restrict

Constructors

TypeQuals

Fields

constant :: Bool

volatile :: Bool

restrict :: Bool

noTypeQuals :: TypeQuals

no type qualifiers

mergeTypeQuals :: TypeQuals -> TypeQuals -> TypeQuals

merge (&&) two type qualifier sets

Variable names

data VarName

VarName name assembler-name is a name of an declared object

Constructors

VarName Ident (Maybe AsmName)

NoName

identOfVarName :: VarName -> Ident

isNoName :: VarName -> Bool

type AsmName = CStrLit

Assembler name (alias for CStrLit)

Attributes (STUB, not yet analyzed)

data Attr

attribute annotations

Those are of the form Attr attribute-name attribute-parameters, and serve as generic properties of some syntax tree elements.

Some examples:

labels can be attributed with unused to indicate that their not used

struct definitions can be attributed with packed to tell the compiler to use the most compact representation

declarations can be attributed with deprecated

function declarations can be attributes with noreturn to tell the compiler that the function will never return,

or with const to indicate that it is a pure function

TODO: ultimatively, we want to parse attributes and represent them in a typed way

Constructors

Attr Ident [Expr] NodeInfo

type Attributes = [Attr]

noAttributes :: Attributes

Empty attribute list

mergeAttributes :: Attributes -> Attributes -> Attributes

Merge attribute lists TODO: currently does not remove duplicates

Statements and Expressions (STUB, aliases to Syntax)

type Stmt = CStat

Stmt is an alias for CStat (Syntax)

type Expr = CExpr

Expr is currently an alias for CExpr (Syntax)

type Initializer = CInit

Initializer is currently an alias for CInit.

We're planning a normalized representation, but this depends on the implementation of constant expression evaluation

type AsmBlock = CStrLit

Top level assembler block (alias for CStrLit)