Language.C.Data.Name

Description

Unique Names with fast equality (newtype Int)

Documentation

newtype Name

Name is a unique identifier

Constructors

Name

Fields

nameId :: Int

newNameSupply :: [Name]

return an infinite stream of Names starting with nameId 0

namesStartingFrom :: Int -> [Name]

get the infinite stream of unique names starting from the given integer