Language.C.Parser

Description

Language.C parser

Simple API

parseC :: InputStream -> Position -> Either ParseError CTranslUnit

parseC input initialPos parses the given preprocessed C-source input and returns the AST or a list of parse errors.

Parser Monad

data P a

execParser :: P a -> InputStream -> Position -> [Ident] -> [Name] -> Either ParseError (a, [Name])

execute the given parser on the supplied input stream. returns ParseError if the parser failed, and a pair of result and remaining name supply otherwise

Synopsis: execParser parser inputStream initialPos predefinedTypedefs uniqNameSupply

execParser_ :: P a -> InputStream -> Position -> Either ParseError a

run the given parser using a new name supply and builtin typedefs see execParser

Synopsis: runParser parser inputStream initialPos

builtinTypeNames :: [Ident]

Exposed Parsers

translUnitP :: P CTranslUnit

translUnitP provides a parser for a complete C translation unit, i.e. a list of external declarations.

extDeclP :: P CExtDecl

extDeclP provides a parser for an external (file-scope) declaration

statementP :: P CStat

statementP provides a parser for C statements

expressionP :: P CExpr

expressionP provides a parser for C expressions

Parser Monad

newtype ParseError

Constructors

ParseError ([String], Position)