Language.C.Pretty

Description

This module provides a pretty printer for the parse tree (AST).

Pretty Printing

class Pretty p where

A class of types which can be pretty printed

Minimal complete definition

Nothing

Methods

pretty :: p -> Doc

pretty print the given value

prettyPrec :: Int -> p -> Doc

prettyPrec prec p pretty prints p assuming that the surrounding context has a precedence of prec

Testing

prettyUsingInclude :: CTranslUnit -> Doc

Pretty print the given tranlation unit, but replace declarations from header files with #include directives.

The resulting file may not compile (because of missing #define directives and similar things), but is very useful for testing, as otherwise the pretty printed file will be cluttered with declarations from system headers.